// src/app/core/api/model/branch.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Tenant } from './tenant';


export interface Branch { 
    id?: number;
    name?: string;
    address?: string;
    phone?: string;
    tenant?: Tenant;
}